part-13-docker-deployment/backend/app/pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.1"
description = "Ultimate FastAPI Tutorial Part 5"
authors = ["ChristopherGS"]

[tool.poetry.dependencies]
python = "^3.8"
uvicorn = "~0.11.3"
fastapi = "~0.68.0"
python-multipart = "~0.0.5"
pydantic = {extras = ["email"], version = "~1.8.1"}
Jinja2 = "~3.0.1"
SQLAlchemy = "~1.4.3"
alembic = "~1.6.5"
tenacity = "~8.0.1"
httpx = "~0.18.1"
passlib = {extras = ["bcrypt"], version = "^1.7.2"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
requests = "^2.26.0"
